#4 Refactoring: Split files to improve readability
