+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
-import { useState, useEffect } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
-import './App.css'
+import { useEffect, useState } from "react";
+import "./App.css";
+import reactLogo from "./assets/react.svg";
+import viteLogo from "/vite.svg";
 
 // 開発環境でのみインポート
 const isDevelopment = import.meta.env.DEV;
@@ -11,40 +11,43 @@ const isDevelopment = import.meta.env.DEV;
 let testsExecuted = false;
 
 function App() {
-  const [count, setCount] = useState(0)
-  const [healthStatus, setHealthStatus] = useState('')
-  const [testStatus, setTestStatus] = useState('Tests not run yet')
+  const [count, setCount] = useState(0);
+  const [healthStatus, setHealthStatus] = useState("");
+  const [testStatus, setTestStatus] = useState("Tests not run yet");
 
   const checkHealth = async () => {
     try {
-      const response = await fetch('http://localhost:8000/api/health')
-      const data = await response.json()
-      setHealthStatus(`API Status: ${data.status}`)
+      const response = await fetch("http://localhost:8000/api/health");
+      const data = await response.json();
+      setHealthStatus(`API Status: ${data.status}`);
     } catch (error) {
-      setHealthStatus('API Error' + error)
+      setHealthStatus("API Error" + error);
     }
-  }
+  };
 
   // 開発環境でIndexedDBテストを自動実行
   // StrictModeによる再マウントでも1回だけ実行されるようグローバル変数を使用
   useEffect(() => {
     if (isDevelopment && !testsExecuted) {
-      testsExecuted = true
-      console.log('Running IndexedDB manual tests in development mode...');
-      setTestStatus('Running tests...');
+      testsExecuted = true;
+      console.log("Running IndexedDB manual tests in development mode...");
+      setTestStatus("Running tests...");
 
       // 動的にテストをインポートして実行
-      import('./db/__tests__/manual-test')
+      import("./__tests__/db/manual-test")
         .then((testModule) => {
           return testModule.runAllTests();
         })
         .then(() => {
-          setTestStatus('✅ Tests completed! Check console for details.');
-          console.log('%cTests completed successfully!', 'color: green; font-weight: bold');
+          setTestStatus("✅ Tests completed! Check console for details.");
+          console.log(
+            "%cTests completed successfully!",
+            "color: green; font-weight: bold"
+          );
         })
         .catch((error) => {
-          setTestStatus('❌ Tests failed! Check console for details.');
-          console.error('Test execution failed:', error);
+          setTestStatus("❌ Tests failed! Check console for details.");
+          console.error("Test execution failed:", error);
         });
     }
   }, []); // 空の依存配列で、マウント時のみ実行
@@ -64,12 +67,17 @@ function App() {
         <button onClick={() => setCount((count) => count + 1)}>
           count is {count}
         </button>
-        <button onClick={checkHealth}>
-          Check API Health
-        </button>
+        <button onClick={checkHealth}>Check API Health</button>
         <p>{healthStatus}</p>
         {isDevelopment && (
-          <div style={{ marginTop: '20px', padding: '10px', background: '#f0f0f0', borderRadius: '5px' }}>
+          <div
+            style={{
+              marginTop: "20px",
+              padding: "10px",
+              background: "#f0f0f0",
+              borderRadius: "5px",
+            }}
+          >
             <strong>IndexedDB Tests:</strong> {testStatus}
           </div>
         )}
@@ -81,7 +89,7 @@ function App() {
         Click on the Vite and React logos to learn more
       </p>
     </>
-  )
+  );
 }
 
-export default App
+export default App;
